Change to TinyDB, extend Database(), create method
Changed the database being used from shelve to TinyDB, and changed the
Database class to a superclass with two new classes (SQL, NoSQL), that
Get() will mediate based in a config file that will be created in the
future.

Refactored the classes and created a few new methods in the process,
but nothing really new.